rename and update action param
